refactor: separate test_deps.ts from deps.ts
