+++ deps.ts
@@ -6,19 +6,6 @@ export {
 export { copyBytes } from "https://deno.land/std@v0.31.0/io/util.ts";
 
 export {
-  assert,
-  assertEquals,
-  assertStrContains,
-  assertThrowsAsync
-} from "https://deno.land/std@v0.31.0/testing/asserts.ts";
-
-export {
-  runTests,
-  test,
-  TestFunction
-} from "https://deno.land/std@v0.31.0/testing/mod.ts";
-
-export {
   Deferred,
   deferred
 } from "https://deno.land/std@v0.31.0/util/async.ts";
