+++ tests/client.ts
@@ -1,4 +1,4 @@
-import { test, assert, assertStrContains } from "../deps.ts";
+import { test, assert, assertStrContains } from "../test_deps.ts";
 import { Client, PostgresError } from "../mod.ts";
 import { TEST_CONNECTION_PARAMS } from "./constants.ts";
