+++ tests/pool.ts
@@ -3,7 +3,7 @@ import {
   assertEquals,
   TestFunction,
   assertThrowsAsync
-} from "../deps.ts";
+} from "../test_deps.ts";
 import { Pool } from "../pool.ts";
 import { delay } from "../utils.ts";
 import { TEST_CONNECTION_PARAMS, DEFAULT_SETUP } from "./constants.ts";
